requirements.txt
Babel==1.3
argparse==1.3.0
astroid==1.3.4
cliff==1.9.0
cmd2==0.6.8
ipython==2.4.1
iso8601==0.1.10
logilab-common==0.63.2
msgpack-python==0.4.5
netaddr==0.7.13
netifaces==0.10.4
oslo.config==1.6.1
oslo.i18n==1.4.0
oslo.serialization==1.3.0
oslo.utils==1.3.0
pbr==0.10.7
prettytable==0.7.2
pylint==1.4.1
pyparsing==2.0.3
python-ceilometerclient==1.0.12
python-keystoneclient==1.1.0
python-neutronclient==2.3.10
python-novaclient==2.20.0
pytz==2014.10
requests==2.5.1
simplejson==3.6.5
six==1.9.0
stevedore==1.2.0
